requirements.txt
jupyter
openai
python-dotenv
mplfinance